+++ SortMerge.hpp
@@ -2,7 +2,7 @@
 #define SORTMERGE_H
 #include <iostream>
 
-void SortMerge(uint32_t* build, uint32_t buildSize, uint32_t* probe, uint32_t probseSize,
+void SortMerge(uint64_t* build, uint32_t buildSize, uint64_t* probe, uint32_t probseSize,
 							 uint32_t numPartitions);
 
 #endif /* SORTMERGE_H */
